Controller framework for inbound products and skus
Initial framework for inbound products and skus. Includes endpoints for
OKL only. Class InboundMessageService manages the bulk of the work, and
delegates to DatabaseService and FlatFileService. Those classes are
just hooks for now.

The framework includes message classes that are responsible for parsing
the underlying messages (either xml or json) and providing attributes
and information to help write the message to the database.