Remove method missing logic from SOLR serializers
The code to handle attributes that did not exist was superseding the
serializer functionality that checks to see if the attribute has been
overridden on the serializer before going to the underlying object to
read the attribute.
https://github.com/rails-api/active_model_serializers/blob/0-10-stable/l
ib/active_model/serializer.rb#L377-L383

This meant that attributes were not available unless explicitly defined
on the serializer. There was no ability to pass through to the object.
This suppressed some valid fields (allow_exposure, units_sold_,
vintage) as well as hid some fields that were incorrectly named
(units_sold_last_week > units_sold_last_1_week, vdc_flag > vdc_sku).

The lists in ProductCoreFields were not changed. If attributes do not
exist for a product, sku, or both, then they are explicitly stubbed.
Ideally this gets cleaned up and there's no need to stub attributes:
either provide a value or remove it from the index.