CAT-821: Update indexer for web flags summary and web status
The following fields were removed from ProductCoreFields:
- web_status_buyer_removed
- web_status_in_progress
- web_status_active
- web_status_suspended

We will no longer be sending values for those fields, and Search API
will be updated so that it does not ask about those fields. The next
time the index is re-created, those fields will go away.

The following field is being re-purposed:
- web_status

It used to hold the web flags summary, but going forward will be
populated with the concept level ADINPU web status.

The following fields were added:
- web_flags_summary
- web_flags_summary_live_on_site
- web_flags_summary_in_workflow
- web_flags_summary_suspended
- web_flags_summary_buyer_reviewed

These fields mimic the old web_status_ fields and will be filled with
an array of concept Ids.