Change decorated attribute mixin to use define_method
With the test in place, it became pretty easy to figure out how to
switch this from class_eval to define_method. Doing so allows the
conversions to string to go away. It also becomes clear that the method
can be changed to receive the field object instead of three of its
properties.