Fix source name defaulting logic for Collection Association
Now that the defaulting logic for source name is different between the
Singular and Collection Association classes, the defaulting needs to
live in the sub-class. it is handled through method default_source_name
instead of the source_name getter. (That has been replaced with
attr_reader.)

The source includes were being determined based on the absence or
presence of the source name, but that was a related fact, not the
driving fact. The driving fact is whether or not the source of the
association is itself. That’s been highlighted by moving it to its own
method that indicates whether or not to nest the source includes.

The tests have been updated to specify the type of association instead
of using the base class.