CAT-1148: Selectively clear margin from CA skus (#194)
If comparison of the US and CA price and cost show that either (1) both
match or (2) both are different then there's no clear evidence that
there is an issue. However, if one matches and one is different (for
example the costs are the same but the prices differ), then it might be
the case that the CA cost and price are not in the same currency. To
mitigate this the margin is cleared in those cases.